Generate structures for an AWS environment and runner
